backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
black==23.12.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
iniconfig==2.3.0
isort==5.13.2
kombu==5.6.0
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
pytest-cov
